eslint omits the test directory
I have no idea what any of that stuff in there is, or what it's meant
to do (it doesn't work), but I do know that it's scattered with lint
errors that make it impossible to get a clean `yarn lint` output. I
toyed with just removing the test directory completely, but on the
offchance that it contains something not auto-generated or otherwise
worthless, I ended up keeping it around but not attempting to lint it.