Changes to maintain status values using ControlledVocab
This requires version ^2.7.3 of stripes-smart-components, in which
`<ControlledVocab>` has been extended to implement
`actuator="refdata"`. The trick here is that the settings page must
itself make a stripes-connect request to determine the ID of the
Refdata category that it wants to manipulate, so it can use that in
the base-URL it passes into `<ControlledVocab>`.