Towards the suppliers display described in PR-238
It's a bit sellotape-and-string at the moment, as much of the
necessary information is missing from the rota entries: we have no
branch _name_, only an AUTHORITY:SYMBOL pair; and no status at all, so
we just make one up.

Also, the link into Directory is pretty vague. At some stage, we'll
want a nice way to properly specify in the UI a link to the entry with
the specified authority and symbol, but for now we ignore the
authority and just use the symbol.