Searching uses filters instead of match/term
Roughly no operational difference at this stage (the purpose of this
change is so we can add further filters). But it turns out that the
old match/term searching was using some kind of fuzzy matching --
maybe just right-truncation -- so we were (incorrectly) finding the
DIKUA directory entry when using the institution code DIKU. That bug
has now been fixed en passant.