Include PALCI logo in pull-slip
At present this is hardwired: the image is right here in the module
source, and the image-size is wired right into the `<img>`
element. Clearly that won't do in the general case, but it will
suffice to get us rolling until the Directory system has proper
support for consortia and can furnish us with an appropriate logo at
run-time.