Include CSS reset before pull-slip CSS
It turns out that Filip wrote the pull-slip CSS on the assumption that
it would be running in a context incuding Eric Meyer's `reset.css`
whereas in fact Stripes uses `normalize.css`. Rather than require
Filip to rewrite the CSS, and risk this coming up again the next time
the stylesheet gets revised, we now explictly include `reset.css`
before the pull-slip styles.

As a result, the pull-slip now looks right!