Fix "close" button for full-record view ... again
This time, instead of using the package-file's `stripes.route` (which
as it turns out is always `/rs` because of course we always get
`ui-rs`'s package file), we use the appName.

Infuriatingly, fetching the appName requires the use of a
`<AppNameContext.Consumer>` which means all the bulk of the code has
to be right-shifted in a whitespace-only delta. So I am here
committing version that works correctly on a minimal delta; and I will
re-indent in a subsequent non-substantive commit.