Remove unnecessary Eslint-disabling comments (#11)
We never needed `eslint-disable-line no-template-curly-in-string` when
using `:{xxx}`, only for `${xxx}` (which we no longer need since
`%{xxx}` is a synonym). I'm not sure how these ESLint-disabling
comments got added, but my best guess is careless copy-pasta from an
over-paranoid module that uses these for _every_ manifest path due to
having once needed it for a `${xxx}`.