Mark slips as printed when printing
Sadly, all we know for sure is that the print dialogue has been
invoked and closed -- but not whether it was closed by printing or by
cancelling. As far as I can tell there is no way for the
`onafterprint` event handler to know this: the event object that's
passed in doesn't seem to contain any indication.