Update user-card layout
Remove patron-group and username, neither of which we are able to
obtain via NCIP, and move email address up into the old username slot.

Fixes PR-500.

These changes render the notion of a reusable "user" card even more
tenuous than previously. Given that, in the context of ReShare, we
will never use either `<SmartUserCard.js>` or `<GraphQLUserCard.js>`,
and given that no other module uses this supposedly reusable code, I
think the honest thing would be to delete those two components, move
the present one to where it's used (in or near
`../../ViewPatronRequest/sections/`) and simplify it to omit the
"general-purpose" support.