When printing all pull-slips, mark them all as printed
This involves inspecting the set of valid actions for each request,
sending actions for those in the `supplierPrintPullSlip` state,
gathering the results of all the resulting promises and reporting one
of four possible states:
* All slips had already been marked as printed
* Protocol-level error (should be a "can't happen", but you know)
* All slips that needed marking were successfully marked
* One or more slips could not be marked for some reason

The last of these should also be a rare condition, but when it does
happen there is potentially a lot of text to be displayed. Since it
appears only in a callout of limited duration, it is also logged to
the console to aid debugging.