Refresh record list after closing the print-all-pull-slips page
This was surprisingly tricky. The reason it doesn't refresh is that
stripes-connect, when fetching the record list, notices that the URL
is the same as it was the last time the resource was fetched, so it
just returns its cached value from that previous fetch. To defeat
this, `<PatronRequestsRoute>` has to inject a dummy additional sort
criterion (which is ignored by mod-rs) so that the URL that is
generated stripes-connect is different when `refresh=1` is part of the
UI URL. And to make _that_ happen, `<PrintAllPullSlips>` has to add
that when it's missing but also remove it when it's present, so that
in either case we end up with the URL change when we return to the
record-list.