Generate date-range searches for creation date
This code presently generates the right UI URLs, and _almost_ the
right back-end URLs; but it fails because the dates coming out of the
`<Datepicker>` are in locale format and not ISO format.

I am not yet sure what to do about that. I can't see a way to have the
`<Datepicker>` emit ISO-format dates, and maybe that it not even
desirable since those dates are displayed for users to see. So perhaps
the right approach is to have the `setFilterDate` function parse those
locale-formatted dates and derive ISO equivalents to set into the UI
URL (whence they will make their way into the back-end URL.)