Datepicker textboxes render correctly on clear
The problem here was caused by a bug in Datepicker where the `value`
prop is ignored except on the first render -- so resetting the value
did not cause the visible value to be cleared.

The "fix" is to use a `key=...` prop in the `<Datepicker>s`, changing
when the value does. This forces a reinstantiation rather than a mere
re-render, so that the current `value` is honoured. Many thanks to
@JohnC-80 for pointing out this workaround. (The alternative would
have been a lot of tedious messing about with React refs to manually
clear the textbox.)

Unfortunately, including these `key` properties causes a benign
console error:
	Warning: Can't perform a React state update on an unmounted component

I hope that in due time we can make this go away just by removing the
`key` properties once the Datepicker rendering bug has been
fixed. With that hope in view, I don't intend to burn time trying to
understand exactly what the cause is.