Work towards maintaining pull-slip notification schedules (#33)
This is not yet complete, but does offer a means to list, view and delete schedules, and also to edit some fields. The editing work is taking longer than expected because it involves so many novel controls (e.g. list of Timepicker). I am merging to master at this point because the work I have been doing has been invisible for so long, but will then continue to work on my branch.