Further attempts to get a label on the list-of-Timepicker
This doesn't really work, as even though we now pass the `label` prop
through to `<RepeatableField>`, it ignores that prop. We work around
it by also putting the value of that prop into `headLabels`. This
"works" in that we do get a label, but (A) it's really the same thing,
and (B) it looks awful because the head-labels are not vertically
aligned with the regular labels of scalar fields in the same `<Row>`.

I hope to presuade John Coburn that `<RepeatableField>` should support
a regular `label`. If that comes off, I can dump the hack that uses
`headLabels`. If not, I don't know where we go from there. The best
visual effect would probably come from using a separate `<Row>` above
the main one, just to contain the labels. But that would be
semantically terrible. We'll see what happens.