Add new permission, `ui-rs.settings.pullslip-notifications`
Part of PR-813.

This permission _should_ be used as a guard on the Pullslip
Notifications settings page. But I can't use it there yet because the
permission can't be added to a user until it exists, and thanks to
MODPERMS-34 it will not exist until the entire ReShare server is torn
down and rebuilt from scratch. *sigh*