Fix legend-alignment problems in pull-slip notification editor
This relies on STCOM-714, which was fixed a week ago. As a result, the
fix is in place when using a tip-of-master stripes-components, but not
with any yet-extant release. But that's OK because the fix is benign:
it just involved passing a new prop which will be ignored by old
releases of stripes components leaving the alignment off exactly as it
was before this fix.

Fixes PR-723.