Create "en.json" file for translations
We were previously using "en_US.json", just because that's what got
set up in the earlist commit and it seemed to work. But now that we're
introducing the use of Lokalize (so non-developers can contribute
translations), we need to use the standard filename "en.json" as the
source.

But we don't -- for now -- want to remove the old file, because we're
not 100% confident that Stripes will back down to directly using "en"
if the "en_US" translation file is absent. So for now, we'll keep
both, and hopefully "en_US.json" will soon start getting overriden by
Lokalize.