Add a dummy "none" option to list of lmsLocations
It's not so much that this is a better representation of what we want
to able to express in a pull-slip notification schedule, as that when
a select only has a single entry (as is presently the case on the East
server), it is impossible to make an actual change to the value of
that select, and that means that no value is set into the submitted
set.

This was the cause of PR-847, which has detained me for far longer
than it ought to have done. Turns out, the reason this never
manifested back when I was first doing the pull-slip notification
schedule settings work is simply that we had more LMS locations in the
server then, so we never ran into the "you've not changed it so I'm
ignoring the value" problem.

So this fixes PR-847.