Omit malformed dates from filters
When the filterstring contains elements beginning with 'date' and not
ending with a well-formed ISO date, those elements are simply
discarded from the generated query.

I had hoped this would fix PR-1002, and it does indeed ensure that the
query sent is always correct and never results in an error-message
from the back-end module.

But that's not good enough, because typing into a date-filter box
still causes a re-render on each character, which defocuses the
date-filter box so that you need to click back into it for each
character.