Format due dates according to locale (#56)
Format due dates according to locale

As discussed in PR-897, the due-date (when specified) can appear in
various different formats, being apparently an uncontrolled text
field. When this is recognizable as an ISO-compatible date string, it
is now interpreted as such and rendered in a local-appropriate
way. When it is not, it is "as is" with a locale-appropriate note that
it's a bad date.

For our purposes, an ISO-compatible date string:

* Begins with YYYY-MM-DD
* _May_ also contain a time part
* The time part, if present, may be separated from the date with `T` or a space.
* If present, the time part must be of the form HH:MM:SS

Fixes PR-1033.