Improve notice template editing
The ContextualisedTemplateForm curried with props and passed to
EntryManager was not memoised and so without a stable reference the form
would be unexpectedly cleared at times.

The initialvalues better reflect the expected structure.

And finally, duplicate templates wouldn't be fully stripped of their ids
nor would it completely avoid mutating the prop, this probably works
better.