Reset query when URL changes rather than whole component
I added a key to the SearchAndSortQuery component to reset when the
search params change in order to enable the link to remove filter in
PR-1228. This commit uses the query setter to do that in a more
targetted way to avoid component state for that whole tree being lost
every time the URL changes.