Fix BackendMultiSelectionFilter to actually debounce and other tidying
Turns out we don't actually need the labels in dataOptions to have them
displayed on the chips so that simplifies things a bit.

useEffect was the wrong thing here, better semantics to have it update
the cache on query success.

The debounce from the asyncFiltering prop didn't actually work
consistently even if we memoize dataOptions and the loading dots were
confusing anyhow so leaving it off for now.

Alas the emptyMessage prop uses useRef so we can't use it to address the
fact it says "No results found" before the search is triggered and while
it's running by setting that.

At least in Ramsons+ the showLoading prop gives a reliable loading
indication.